qwertyuiop/requirements.txt
# Core dependencies
asyncio
websockets>=11.0.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
librosa>=0.10.0

# UltraVAD dependencies
transformers[torch]>=4.30.0

# Audio processing
piper-tts>=1.2.0

# Additional utilities
uuid
wave
tempfile
io
json
logging
typing
concurrent.futures
